backend/requirements.txt
flask
newspaper3k
google
google-genai
python-dotenv
lxml_html_clean
gunicorn